--- aperag/migration/versions/20250730110318-f7671d2a7bff.py
@@ -1,39 +1,0 @@
-"""empty message
-
-Revision ID: f7671d2a7bff
-Revises: b598e645b2ba
-Create Date: 2025-07-30 11:03:18.975296
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f7671d2a7bff'
-down_revision: Union[str, None] = 'b598e645b2ba'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('setting',
-    sa.Column('key', sa.String(length=256), nullable=False),
-    sa.Column('value', sa.Text(), nullable=True),
-    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
-    sa.PrimaryKeyConstraint('key')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('setting')
-    # ### end Alembic commands ###
