--- aperag/migration/versions/20250805104914-5b7f5b4b3765.py
@@ -1,36 +1,0 @@
-"""empty message
-
-Revision ID: 5b7f5b4b3765
-Revises: 69b1deed0316
-Create Date: 2025-08-05 10:49:14.015449
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5b7f5b4b3765'
-down_revision: Union[str, None] = '69b1deed0316'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user_quota', sa.Column('quota_limit', sa.Integer(), nullable=False))
-    op.add_column('user_quota', sa.Column('current_usage', sa.Integer(), nullable=False))
-    op.drop_column('user_quota', 'value')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user_quota', sa.Column('value', sa.INTEGER(), autoincrement=False, nullable=False))
-    op.drop_column('user_quota', 'current_usage')
-    op.drop_column('user_quota', 'quota_limit')
-    # ### end Alembic commands ###
