--- aperag/migration/versions/20250806163353-f53af2d39508.py
@@ -1,53 +1,0 @@
-"""empty message
-
-Revision ID: f53af2d39508
-Revises: 5b7f5b4b3765
-Create Date: 2025-08-06 16:33:53.156828
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f53af2d39508'
-down_revision: Union[str, None] = '5b7f5b4b3765'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('oauth_account',
-    sa.Column('id', sa.String(length=24), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('oauth_name', sa.String(length=100), nullable=False),
-    sa.Column('access_token', sa.String(length=1024), nullable=False),
-    sa.Column('expires_at', sa.Integer(), nullable=True),
-    sa.Column('refresh_token', sa.String(length=1024), nullable=True),
-    sa.Column('account_id', sa.String(length=320), nullable=False),
-    sa.Column('account_email', sa.String(length=320), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ondelete='cascade'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_oauth_account_account_id'), 'oauth_account', ['account_id'], unique=False)
-    op.create_index(op.f('ix_oauth_account_oauth_name'), 'oauth_account', ['oauth_name'], unique=False)
-    op.alter_column('user', 'username',
-               existing_type=sa.VARCHAR(length=256),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'username',
-               existing_type=sa.VARCHAR(length=256),
-               nullable=False)
-    op.drop_index(op.f('ix_oauth_account_oauth_name'), table_name='oauth_account')
-    op.drop_index(op.f('ix_oauth_account_account_id'), table_name='oauth_account')
-    op.drop_table('oauth_account')
-    # ### end Alembic commands ###
