--- aperag/migration/versions/20250825112501-6622089995c7.py
@@ -1,34 +1,0 @@
-"""empty message
-
-Revision ID: 6622089995c7
-Revises: 413ebe1ef0d5
-Create Date: 2025-08-25 11:25:01.720822
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '6622089995c7'
-down_revision: Union[str, None] = '413ebe1ef0d5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_message_feedback_collection_id'), table_name='message_feedback')
-    op.drop_column('message_feedback', 'collection_id')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('message_feedback', sa.Column('collection_id', sa.VARCHAR(length=24), autoincrement=False, nullable=True))
-    op.create_index(op.f('ix_message_feedback_collection_id'), 'message_feedback', ['collection_id'], unique=False)
-    # ### end Alembic commands ###
