--- aperag/migration/versions/20250826114238-df7a2c40ed6e.py
@@ -1,34 +1,0 @@
-"""empty message
-
-Revision ID: df7a2c40ed6e
-Revises: 6622089995c7
-Create Date: 2025-08-26 11:42:38.715839
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'df7a2c40ed6e'
-down_revision: Union[str, None] = '6622089995c7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('document', sa.Column('content_hash', sa.String(length=64), nullable=True))
-    op.create_index(op.f('ix_document_content_hash'), 'document', ['content_hash'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_document_content_hash'), table_name='document')
-    op.drop_column('document', 'content_hash')
-    # ### end Alembic commands ###
