--- aperag/migration/versions/20250829101722-624ebebacb56.py
@@ -1,34 +1,0 @@
-"""empty message
-
-Revision ID: 624ebebacb56
-Revises: df7a2c40ed6e
-Create Date: 2025-08-29 10:17:22.611888
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '624ebebacb56'
-down_revision: Union[str, None] = 'df7a2c40ed6e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('chat_collection_id', sa.String(length=24), nullable=True))
-    op.create_index(op.f('ix_user_chat_collection_id'), 'user', ['chat_collection_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_user_chat_collection_id'), table_name='user')
-    op.drop_column('user', 'chat_collection_id')
-    # ### end Alembic commands ###
