--- aperag/migration/versions/20250903145115-f12511e5abef.py
@@ -1,36 +1,0 @@
-"""empty message
-
-Revision ID: f12511e5abef
-Revises: 624ebebacb56
-Create Date: 2025-09-03 14:51:15.253268
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f12511e5abef'
-down_revision: Union[str, None] = '624ebebacb56'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_index('idx_evaluations_collection_id', 'evaluations', ['collection_id'], unique=False)
-    op.add_column('question_sets', sa.Column('collection_id', sa.String(length=24), nullable=True))
-    op.create_index('idx_question_sets_collection_id', 'question_sets', ['collection_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('idx_question_sets_collection_id', table_name='question_sets')
-    op.drop_column('question_sets', 'collection_id')
-    op.drop_index('idx_evaluations_collection_id', table_name='evaluations')
-    # ### end Alembic commands ###
