--- aperag/migration/versions/20250904181845-d3f3e2ea757c.py
@@ -1,40 +1,0 @@
-"""empty message
-
-Revision ID: d3f3e2ea757c
-Revises: f12511e5abef
-Create Date: 2025-09-04 18:18:45.143661
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'd3f3e2ea757c'
-down_revision: Union[str, None] = 'f12511e5abef'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('bot_collection_relation')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('bot_collection_relation',
-    sa.Column('id', sa.VARCHAR(length=24), autoincrement=False, nullable=False),
-    sa.Column('bot_id', sa.VARCHAR(length=24), autoincrement=False, nullable=False),
-    sa.Column('collection_id', sa.VARCHAR(length=24), autoincrement=False, nullable=False),
-    sa.Column('gmt_created', postgresql.TIMESTAMP(timezone=True), autoincrement=False, nullable=False),
-    sa.Column('gmt_deleted', postgresql.TIMESTAMP(timezone=True), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name=op.f('bot_collection_relation_pkey')),
-    sa.UniqueConstraint('bot_id', 'collection_id', 'gmt_deleted', name=op.f('uq_bot_collection_deleted'))
-    )
-    # ### end Alembic commands ###
