--- aperag/migration/versions/20250905212643-d112e0332219.py
@@ -1,32 +1,0 @@
-"""empty message
-
-Revision ID: d112e0332219
-Revises: d3f3e2ea757c
-Create Date: 2025-09-05 21:26:43.965801
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd112e0332219'
-down_revision: Union[str, None] = 'd3f3e2ea757c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('searchhistory', sa.Column('vision_search', sa.JSON(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('searchhistory', 'vision_search')
-    # ### end Alembic commands ###
