+++ aperag/migration/versions/20260425041904-6a8fc31427d9.py
@@ -1,0 +1,865 @@
+"""post_phase_7_initial
+
+Revision ID: 6a8fc31427d9
+Revises: 
+Create Date: 2026-04-25 04:19:04.615161
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '6a8fc31427d9'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('agent_artifact',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('turn_id', sa.String(length=24), nullable=False),
+    sa.Column('artifact_type', sa.String(length=50), nullable=False),
+    sa.Column('summary', sa.Text(), nullable=True),
+    sa.Column('payload', sa.JSON(), nullable=False),
+    sa.Column('storage_ref', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_agent_artifact_turn_type', 'agent_artifact', ['turn_id', 'artifact_type'], unique=False)
+    op.create_index(op.f('ix_agent_artifact_artifact_type'), 'agent_artifact', ['artifact_type'], unique=False)
+    op.create_index(op.f('ix_agent_artifact_turn_id'), 'agent_artifact', ['turn_id'], unique=False)
+    op.create_table('agent_timeline_event',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('turn_id', sa.String(length=24), nullable=False),
+    sa.Column('sequence', sa.Integer(), nullable=False),
+    sa.Column('timestamp', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('type', sa.String(length=128), nullable=False),
+    sa.Column('label', sa.String(length=128), nullable=True),
+    sa.Column('status', sa.String(length=64), nullable=True),
+    sa.Column('actor', sa.String(length=50), nullable=False),
+    sa.Column('data', sa.JSON(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('turn_id', 'sequence', name='uq_agent_timeline_event_turn_sequence')
+    )
+    op.create_index('idx_agent_timeline_event_turn_timestamp', 'agent_timeline_event', ['turn_id', 'timestamp'], unique=False)
+    op.create_index(op.f('ix_agent_timeline_event_timestamp'), 'agent_timeline_event', ['timestamp'], unique=False)
+    op.create_index(op.f('ix_agent_timeline_event_turn_id'), 'agent_timeline_event', ['turn_id'], unique=False)
+    op.create_index(op.f('ix_agent_timeline_event_type'), 'agent_timeline_event', ['type'], unique=False)
+    op.create_table('agent_turn',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('chat_id', sa.String(length=24), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('bot_id', sa.String(length=24), nullable=False),
+    sa.Column('request_id', sa.String(length=64), nullable=False),
+    sa.Column('client_idempotency_key', sa.String(length=128), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('input_text', sa.Text(), nullable=False),
+    sa.Column('model_profile', sa.JSON(), nullable=False),
+    sa.Column('error_code', sa.String(length=128), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('answer_artifact_id', sa.String(length=24), nullable=True),
+    sa.Column('reference_bundle_artifact_id', sa.String(length=24), nullable=True),
+    sa.Column('timeline_cursor', sa.Integer(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_started', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_finished', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('chat_id', 'client_idempotency_key', name='uq_agent_turn_chat_idempotency')
+    )
+    op.create_index('idx_agent_turn_chat_created', 'agent_turn', ['chat_id', 'gmt_created'], unique=False)
+    op.create_index('idx_agent_turn_user_status', 'agent_turn', ['user', 'status'], unique=False)
+    op.create_index(op.f('ix_agent_turn_answer_artifact_id'), 'agent_turn', ['answer_artifact_id'], unique=False)
+    op.create_index(op.f('ix_agent_turn_bot_id'), 'agent_turn', ['bot_id'], unique=False)
+    op.create_index(op.f('ix_agent_turn_chat_id'), 'agent_turn', ['chat_id'], unique=False)
+    op.create_index(op.f('ix_agent_turn_reference_bundle_artifact_id'), 'agent_turn', ['reference_bundle_artifact_id'], unique=False)
+    op.create_index(op.f('ix_agent_turn_request_id'), 'agent_turn', ['request_id'], unique=True)
+    op.create_index(op.f('ix_agent_turn_status'), 'agent_turn', ['status'], unique=False)
+    op.create_index(op.f('ix_agent_turn_user'), 'agent_turn', ['user'], unique=False)
+    op.create_table('api_key',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('key', sa.String(length=64), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('description', sa.String(length=256), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('is_system', sa.Boolean(), nullable=False),
+    sa.Column('last_used_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_api_key_gmt_deleted'), 'api_key', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_api_key_is_system'), 'api_key', ['is_system'], unique=False)
+    op.create_index(op.f('ix_api_key_status'), 'api_key', ['status'], unique=False)
+    op.create_index(op.f('ix_api_key_user'), 'api_key', ['user'], unique=False)
+    op.create_table('audit_log',
+    sa.Column('id', sa.String(length=36), nullable=False),
+    sa.Column('user_id', sa.String(length=36), nullable=True, comment='User ID'),
+    sa.Column('username', sa.String(length=255), nullable=True, comment='Username'),
+    sa.Column('resource_type', sa.String(length=50), nullable=True, comment='Resource type'),
+    sa.Column('resource_id', sa.String(length=255), nullable=True, comment='Resource ID (extracted at query time)'),
+    sa.Column('api_name', sa.String(length=255), nullable=False, comment='API operation name'),
+    sa.Column('http_method', sa.String(length=10), nullable=False, comment='HTTP method (POST, PUT, DELETE)'),
+    sa.Column('path', sa.String(length=512), nullable=False, comment='API path'),
+    sa.Column('status_code', sa.Integer(), nullable=True, comment='HTTP status code'),
+    sa.Column('request_data', sa.Text(), nullable=True, comment='Request data (JSON)'),
+    sa.Column('response_data', sa.Text(), nullable=True, comment='Response data (JSON)'),
+    sa.Column('error_message', sa.Text(), nullable=True, comment='Error message if failed'),
+    sa.Column('ip_address', sa.String(length=45), nullable=True, comment='Client IP address'),
+    sa.Column('user_agent', sa.String(length=500), nullable=True, comment='User agent string'),
+    sa.Column('request_id', sa.String(length=255), nullable=False, comment='Request ID for tracking'),
+    sa.Column('start_time', sa.BigInteger(), nullable=False, comment='Request start time (milliseconds since epoch)'),
+    sa.Column('end_time', sa.BigInteger(), nullable=True, comment='Request end time (milliseconds since epoch)'),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False, comment='Created time'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_audit_api_name', 'audit_log', ['api_name'], unique=False)
+    op.create_index('idx_audit_gmt_created', 'audit_log', ['gmt_created'], unique=False)
+    op.create_index('idx_audit_http_method', 'audit_log', ['http_method'], unique=False)
+    op.create_index('idx_audit_request_id', 'audit_log', ['request_id'], unique=False)
+    op.create_index('idx_audit_resource_id', 'audit_log', ['resource_id'], unique=False)
+    op.create_index('idx_audit_resource_type', 'audit_log', ['resource_type'], unique=False)
+    op.create_index('idx_audit_start_time', 'audit_log', ['start_time'], unique=False)
+    op.create_index('idx_audit_status_code', 'audit_log', ['status_code'], unique=False)
+    op.create_index('idx_audit_user_id', 'audit_log', ['user_id'], unique=False)
+    op.create_table('bot',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('title', sa.String(length=256), nullable=True),
+    sa.Column('type', sa.String(length=50), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('config', sa.Text(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_bot_gmt_deleted'), 'bot', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_bot_status'), 'bot', ['status'], unique=False)
+    op.create_index(op.f('ix_bot_user'), 'bot', ['user'], unique=False)
+    op.create_table('chat',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('peer_type', sa.String(length=50), nullable=False),
+    sa.Column('peer_id', sa.String(length=256), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('bot_id', sa.String(length=24), nullable=False),
+    sa.Column('title', sa.String(length=256), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('bot_id', 'peer_type', 'peer_id', 'gmt_deleted', name='uq_chat_bot_peer_deleted')
+    )
+    op.create_index(op.f('ix_chat_bot_id'), 'chat', ['bot_id'], unique=False)
+    op.create_index(op.f('ix_chat_gmt_deleted'), 'chat', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_chat_status'), 'chat', ['status'], unique=False)
+    op.create_index(op.f('ix_chat_user'), 'chat', ['user'], unique=False)
+    op.create_table('collection',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('title', sa.String(length=256), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('type', sa.String(length=50), nullable=False),
+    sa.Column('config', sa.Text(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_collection_gmt_deleted'), 'collection', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_collection_status'), 'collection', ['status'], unique=False)
+    op.create_index(op.f('ix_collection_user'), 'collection', ['user'], unique=False)
+    op.create_table('collection_marketplace',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=False),
+    sa.Column('status', sa.String(length=20), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('collection_id', name='uq_collection_marketplace_collection')
+    )
+    op.create_index('idx_collection_marketplace_collection_id', 'collection_marketplace', ['collection_id'], unique=False)
+    op.create_index('idx_collection_marketplace_gmt_deleted', 'collection_marketplace', ['gmt_deleted'], unique=False)
+    op.create_index('idx_collection_marketplace_list', 'collection_marketplace', ['status', 'gmt_created'], unique=False)
+    op.create_index('idx_collection_marketplace_status', 'collection_marketplace', ['status'], unique=False)
+    op.create_table('collection_summary',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('version', sa.Integer(), nullable=False),
+    sa.Column('observed_version', sa.Integer(), nullable=False),
+    sa.Column('summary', sa.Text(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('processing_token', sa.String(length=64), nullable=True),
+    sa.Column('lease_expires_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_last_reconciled', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('collection_id', name='uq_collection_summary')
+    )
+    op.create_index('idx_collection_summary_status_lease', 'collection_summary', ['status', 'lease_expires_at'], unique=False)
+    op.create_index(op.f('ix_collection_summary_collection_id'), 'collection_summary', ['collection_id'], unique=False)
+    op.create_index(op.f('ix_collection_summary_status'), 'collection_summary', ['status'], unique=False)
+    op.create_table('config',
+    sa.Column('key', sa.String(length=256), nullable=False),
+    sa.Column('value', sa.Text(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('key')
+    )
+    op.create_table('document',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('name', sa.String(length=1024), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('size', sa.BigInteger(), nullable=False),
+    sa.Column('content_hash', sa.String(length=64), nullable=True),
+    sa.Column('object_path', sa.Text(), nullable=True),
+    sa.Column('doc_metadata', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_document_collection_id'), 'document', ['collection_id'], unique=False)
+    op.create_index(op.f('ix_document_content_hash'), 'document', ['content_hash'], unique=False)
+    op.create_index(op.f('ix_document_gmt_deleted'), 'document', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_document_status'), 'document', ['status'], unique=False)
+    op.create_index(op.f('ix_document_user'), 'document', ['user'], unique=False)
+    op.create_index('uq_document_collection_name_active', 'document', ['collection_id', 'name'], unique=True, postgresql_where=sa.text('gmt_deleted IS NULL'))
+    op.create_table('document_index',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('document_id', sa.String(length=24), nullable=False),
+    sa.Column('index_type', sa.String(length=50), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('version', sa.Integer(), nullable=False),
+    sa.Column('observed_version', sa.Integer(), nullable=False),
+    sa.Column('index_data', sa.Text(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('processing_token', sa.String(length=64), nullable=True),
+    sa.Column('lease_expires_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_last_reconciled', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('document_id', 'index_type', name='uq_document_index')
+    )
+    op.create_index('idx_document_index_status_lease', 'document_index', ['status', 'lease_expires_at'], unique=False)
+    op.create_index(op.f('ix_document_index_document_id'), 'document_index', ['document_id'], unique=False)
+    op.create_index(op.f('ix_document_index_id'), 'document_index', ['id'], unique=False)
+    op.create_index(op.f('ix_document_index_index_type'), 'document_index', ['index_type'], unique=False)
+    op.create_index(op.f('ix_document_index_status'), 'document_index', ['status'], unique=False)
+    op.create_table('evaluation_dataset_items',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('dataset_id', sa.String(length=32), nullable=False),
+    sa.Column('case_key', sa.String(length=128), nullable=False),
+    sa.Column('input_message', sa.Text(), nullable=False),
+    sa.Column('expected_answer', sa.Text(), nullable=True),
+    sa.Column('reference_context', sa.Text(), nullable=True),
+    sa.Column('tags', sa.JSON(), nullable=True),
+    sa.Column('case_metadata', sa.JSON(), nullable=True),
+    sa.Column('sort_key', sa.Integer(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('dataset_id', 'case_key', name='uq_evaluation_dataset_item_case_key')
+    )
+    op.create_index('idx_evaluation_dataset_items_dataset', 'evaluation_dataset_items', ['dataset_id'], unique=False)
+    op.create_index(op.f('ix_evaluation_dataset_items_gmt_deleted'), 'evaluation_dataset_items', ['gmt_deleted'], unique=False)
+    op.create_table('evaluation_datasets',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('user_id', sa.String(length=256), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=True),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('source_type', sa.String(length=50), nullable=False),
+    sa.Column('schema_hint', sa.JSON(), nullable=True),
+    sa.Column('item_count', sa.Integer(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_evaluation_datasets_collection', 'evaluation_datasets', ['collection_id'], unique=False)
+    op.create_index('idx_evaluation_datasets_user', 'evaluation_datasets', ['user_id'], unique=False)
+    op.create_index(op.f('ix_evaluation_datasets_gmt_deleted'), 'evaluation_datasets', ['gmt_deleted'], unique=False)
+    op.create_table('evaluation_items',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('evaluation_id', sa.String(length=24), nullable=False),
+    sa.Column('question_id', sa.String(length=24), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('question_text', sa.Text(), nullable=False),
+    sa.Column('ground_truth', sa.Text(), nullable=False),
+    sa.Column('rag_answer', sa.Text(), nullable=True),
+    sa.Column('rag_answer_details', sa.JSON(), nullable=True),
+    sa.Column('llm_judge_score', sa.Integer(), nullable=True),
+    sa.Column('llm_judge_reasoning', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_evaluation_items_evaluation_id', 'evaluation_items', ['evaluation_id'], unique=False)
+    op.create_index(op.f('ix_evaluation_items_status'), 'evaluation_items', ['status'], unique=False)
+    op.create_table('evaluation_run_item_attempts',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('run_item_id', sa.String(length=32), nullable=False),
+    sa.Column('run_id', sa.String(length=32), nullable=False),
+    sa.Column('attempt_no', sa.Integer(), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('agent_chat_id', sa.String(length=24), nullable=True),
+    sa.Column('agent_turn_id', sa.String(length=24), nullable=True),
+    sa.Column('answer_text', sa.Text(), nullable=True),
+    sa.Column('judge_result', sa.JSON(), nullable=True),
+    sa.Column('score', sa.Numeric(precision=6, scale=3), nullable=True),
+    sa.Column('latency_ms', sa.Integer(), nullable=True),
+    sa.Column('token_usage', sa.JSON(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('retry_reason', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_started', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_finished', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('run_item_id', 'attempt_no', name='uq_evaluation_run_item_attempt_no')
+    )
+    op.create_index('idx_evaluation_run_item_attempts_item', 'evaluation_run_item_attempts', ['run_item_id'], unique=False)
+    op.create_index('idx_evaluation_run_item_attempts_run', 'evaluation_run_item_attempts', ['run_id'], unique=False)
+    op.create_table('evaluation_run_items',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('run_id', sa.String(length=32), nullable=False),
+    sa.Column('source_dataset_item_id', sa.String(length=32), nullable=True),
+    sa.Column('case_key', sa.String(length=128), nullable=False),
+    sa.Column('sort_key', sa.Integer(), nullable=False),
+    sa.Column('input_message', sa.Text(), nullable=False),
+    sa.Column('expected_answer', sa.Text(), nullable=True),
+    sa.Column('reference_context', sa.Text(), nullable=True),
+    sa.Column('tags', sa.JSON(), nullable=True),
+    sa.Column('case_metadata', sa.JSON(), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('best_score', sa.Numeric(precision=6, scale=3), nullable=True),
+    sa.Column('latest_attempt_id', sa.String(length=32), nullable=True),
+    sa.Column('attempt_count', sa.Integer(), nullable=False),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_evaluation_run_items_run', 'evaluation_run_items', ['run_id'], unique=False)
+    op.create_index('idx_evaluation_run_items_status', 'evaluation_run_items', ['status'], unique=False)
+    op.create_table('evaluation_runs',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('user_id', sa.String(length=256), nullable=False),
+    sa.Column('bot_id', sa.String(length=24), nullable=False),
+    sa.Column('dataset_id', sa.String(length=32), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=True),
+    sa.Column('dataset_name', sa.String(length=255), nullable=True),
+    sa.Column('name', sa.String(length=255), nullable=True),
+    sa.Column('bot_config_snapshot', sa.JSON(), nullable=True),
+    sa.Column('model_config_snapshot', sa.JSON(), nullable=True),
+    sa.Column('judge_config', sa.JSON(), nullable=True),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('summary', sa.JSON(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_started', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_finished', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_evaluation_runs_bot', 'evaluation_runs', ['bot_id'], unique=False)
+    op.create_index('idx_evaluation_runs_collection', 'evaluation_runs', ['collection_id'], unique=False)
+    op.create_index('idx_evaluation_runs_dataset', 'evaluation_runs', ['dataset_id'], unique=False)
+    op.create_index('idx_evaluation_runs_status', 'evaluation_runs', ['status'], unique=False)
+    op.create_index('idx_evaluation_runs_user', 'evaluation_runs', ['user_id'], unique=False)
+    op.create_table('evaluations',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user_id', sa.String(length=24), nullable=False),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=False),
+    sa.Column('question_set_id', sa.String(length=24), nullable=False),
+    sa.Column('agent_llm_config', sa.JSON(), nullable=False),
+    sa.Column('judge_llm_config', sa.JSON(), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('total_questions', sa.Integer(), nullable=False),
+    sa.Column('completed_questions', sa.Integer(), nullable=False),
+    sa.Column('average_score', sa.Numeric(precision=3, scale=2), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_evaluations_collection_id', 'evaluations', ['collection_id'], unique=False)
+    op.create_index('idx_evaluations_status', 'evaluations', ['status'], unique=False)
+    op.create_index('idx_evaluations_user_id', 'evaluations', ['user_id'], unique=False)
+    op.create_table('export_task',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('progress', sa.Integer(), nullable=True),
+    sa.Column('message', sa.Text(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('object_store_path', sa.Text(), nullable=True),
+    sa.Column('file_size', sa.BigInteger(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_completed', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_expires', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_export_task_expires', 'export_task', ['status', 'gmt_expires'], unique=False)
+    op.create_index('idx_export_task_user_status', 'export_task', ['user', 'status'], unique=False)
+    op.create_index(op.f('ix_export_task_collection_id'), 'export_task', ['collection_id'], unique=False)
+    op.create_index(op.f('ix_export_task_user'), 'export_task', ['user'], unique=False)
+    op.create_table('graph_curation_runs',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('user_id', sa.String(length=256), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('config_json', sa.JSON(), nullable=True),
+    sa.Column('stats', sa.JSON(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_started', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('gmt_finished', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_graph_curation_runs_collection_status', 'graph_curation_runs', ['collection_id', 'status'], unique=False)
+    op.create_index('idx_graph_curation_runs_user_created', 'graph_curation_runs', ['user_id', 'gmt_created'], unique=False)
+    op.create_table('graph_curation_suggestions',
+    sa.Column('id', sa.String(length=32), nullable=False),
+    sa.Column('run_id', sa.String(length=32), nullable=False),
+    sa.Column('user_id', sa.String(length=256), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('entity_ids', sa.JSON(), nullable=False),
+    sa.Column('entity_snapshots', sa.JSON(), nullable=False),
+    sa.Column('target_entity_id', sa.String(length=255), nullable=False),
+    sa.Column('confidence_score', sa.Numeric(precision=6, scale=3), nullable=False),
+    sa.Column('reason', sa.Text(), nullable=False),
+    sa.Column('evidence', sa.JSON(), nullable=True),
+    sa.Column('resolution_note', sa.Text(), nullable=True),
+    sa.Column('operated_by', sa.String(length=256), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_operated', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_graph_curation_suggestions_collection_created', 'graph_curation_suggestions', ['collection_id', 'gmt_created'], unique=False)
+    op.create_index('idx_graph_curation_suggestions_collection_status', 'graph_curation_suggestions', ['collection_id', 'status'], unique=False)
+    op.create_index('idx_graph_curation_suggestions_run', 'graph_curation_suggestions', ['run_id'], unique=False)
+    op.create_table('graphindex_chunks',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('collection_id', sa.String(length=255), nullable=False),
+    sa.Column('chunk_id', sa.String(length=64), nullable=False),
+    sa.Column('doc_id', sa.String(length=255), nullable=False),
+    sa.Column('order_in_doc', sa.Integer(), server_default='0', nullable=False),
+    sa.Column('text', sa.Text(), nullable=False),
+    sa.Column('file_path', sa.Text(), server_default='', nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('collection_id', 'chunk_id', name='uq_graphindex_chunks_collection_chunk')
+    )
+    op.create_index('idx_graphindex_chunks_collection_doc', 'graphindex_chunks', ['collection_id', 'doc_id'], unique=False)
+    op.create_table('graphindex_edges',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('collection_id', sa.String(length=255), nullable=False),
+    sa.Column('source_id', sa.String(length=255), nullable=False),
+    sa.Column('target_id', sa.String(length=255), nullable=False),
+    sa.Column('description', sa.Text(), server_default='', nullable=False),
+    sa.Column('weight', sa.Numeric(precision=6, scale=3), server_default='0', nullable=False),
+    sa.Column('source_chunk_ids', sa.ARRAY(sa.String()), server_default='{}', nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('collection_id', 'source_id', 'target_id', name='uq_graphindex_edges_collection_src_tgt')
+    )
+    op.create_index('idx_graphindex_edges_chunks', 'graphindex_edges', ['source_chunk_ids'], unique=False, postgresql_using='gin')
+    op.create_index('idx_graphindex_edges_collection_src', 'graphindex_edges', ['collection_id', 'source_id'], unique=False)
+    op.create_index('idx_graphindex_edges_collection_tgt', 'graphindex_edges', ['collection_id', 'target_id'], unique=False)
+    op.create_table('graphindex_nodes',
+    sa.Column('id', sa.BigInteger(), autoincrement=True, nullable=False),
+    sa.Column('collection_id', sa.String(length=255), nullable=False),
+    sa.Column('entity_id', sa.String(length=255), nullable=False),
+    sa.Column('name', sa.Text(), nullable=False),
+    sa.Column('type', sa.String(length=64), nullable=False),
+    sa.Column('description', sa.Text(), server_default='', nullable=False),
+    sa.Column('source_chunk_ids', sa.ARRAY(sa.String()), server_default='{}', nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('collection_id', 'entity_id', name='uq_graphindex_nodes_collection_entity')
+    )
+    op.create_index('idx_graphindex_nodes_chunks', 'graphindex_nodes', ['source_chunk_ids'], unique=False, postgresql_using='gin')
+    op.create_index('idx_graphindex_nodes_collection_name', 'graphindex_nodes', ['collection_id', 'name'], unique=False)
+    op.create_index('idx_graphindex_nodes_collection_type', 'graphindex_nodes', ['collection_id', 'type'], unique=False)
+    op.create_table('invitation',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('email', sa.String(length=254), nullable=False),
+    sa.Column('token', sa.String(length=64), nullable=False),
+    sa.Column('created_by', sa.String(length=256), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('expires_at', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('is_used', sa.Boolean(), nullable=False),
+    sa.Column('used_at', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('role', sa.String(length=50), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('token')
+    )
+    op.create_table('llm_provider',
+    sa.Column('name', sa.String(length=128), nullable=False),
+    sa.Column('user_id', sa.String(length=256), nullable=False),
+    sa.Column('label', sa.String(length=256), nullable=False),
+    sa.Column('completion_dialect', sa.String(length=64), nullable=False),
+    sa.Column('embedding_dialect', sa.String(length=64), nullable=False),
+    sa.Column('rerank_dialect', sa.String(length=64), nullable=False),
+    sa.Column('allow_custom_base_url', sa.Boolean(), nullable=False),
+    sa.Column('base_url', sa.String(length=512), nullable=False),
+    sa.Column('extra', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('name')
+    )
+    op.create_index(op.f('ix_llm_provider_user_id'), 'llm_provider', ['user_id'], unique=False)
+    op.create_table('llm_provider_models',
+    sa.Column('provider_name', sa.String(length=128), nullable=False),
+    sa.Column('api', sa.String(length=50), nullable=False),
+    sa.Column('model', sa.String(length=256), nullable=False),
+    sa.Column('custom_llm_provider', sa.String(length=128), nullable=False),
+    sa.Column('context_window', sa.Integer(), nullable=True),
+    sa.Column('max_input_tokens', sa.Integer(), nullable=True),
+    sa.Column('max_output_tokens', sa.Integer(), nullable=True),
+    sa.Column('tags', sa.JSON(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('provider_name', 'api', 'model')
+    )
+    op.create_table('model_service_provider',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('name', sa.String(length=256), nullable=False),
+    sa.Column('status', sa.String(length=50), nullable=False),
+    sa.Column('api_key', sa.String(length=256), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('name', 'gmt_deleted', name='uq_model_service_provider_name_deleted')
+    )
+    op.create_index(op.f('ix_model_service_provider_gmt_deleted'), 'model_service_provider', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_model_service_provider_name'), 'model_service_provider', ['name'], unique=False)
+    op.create_index(op.f('ix_model_service_provider_status'), 'model_service_provider', ['status'], unique=False)
+    op.create_table('prompt_template',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('prompt_type', sa.String(length=50), nullable=False),
+    sa.Column('scope', sa.String(length=20), nullable=False),
+    sa.Column('user_id', sa.String(length=256), nullable=True),
+    sa.Column('content', sa.Text(), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_prompt_template_gmt_deleted'), 'prompt_template', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_prompt_template_prompt_type'), 'prompt_template', ['prompt_type'], unique=False)
+    op.create_index(op.f('ix_prompt_template_scope'), 'prompt_template', ['scope'], unique=False)
+    op.create_index(op.f('ix_prompt_template_user_id'), 'prompt_template', ['user_id'], unique=False)
+    op.create_table('question_sets',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user_id', sa.String(length=24), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=True),
+    sa.Column('name', sa.String(length=255), nullable=False),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_question_sets_collection_id', 'question_sets', ['collection_id'], unique=False)
+    op.create_index('idx_question_sets_user_id', 'question_sets', ['user_id'], unique=False)
+    op.create_table('questions',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('question_set_id', sa.String(length=24), nullable=False),
+    sa.Column('question_type', sa.String(length=50), nullable=True),
+    sa.Column('question_text', sa.Text(), nullable=False),
+    sa.Column('ground_truth', sa.Text(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index('idx_questions_question_set_id', 'questions', ['question_set_id'], unique=False)
+    op.create_table('searchhistory',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('collection_id', sa.String(length=24), nullable=True),
+    sa.Column('query', sa.Text(), nullable=False),
+    sa.Column('vector_search', sa.JSON(), nullable=True),
+    sa.Column('fulltext_search', sa.JSON(), nullable=True),
+    sa.Column('graph_search', sa.JSON(), nullable=True),
+    sa.Column('summary_search', sa.JSON(), nullable=True),
+    sa.Column('vision_search', sa.JSON(), nullable=True),
+    sa.Column('items', sa.JSON(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_searchhistory_collection_id'), 'searchhistory', ['collection_id'], unique=False)
+    op.create_index(op.f('ix_searchhistory_gmt_deleted'), 'searchhistory', ['gmt_deleted'], unique=False)
+    op.create_index(op.f('ix_searchhistory_user'), 'searchhistory', ['user'], unique=False)
+    op.create_table('setting',
+    sa.Column('key', sa.String(length=256), nullable=False),
+    sa.Column('value', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('key')
+    )
+    op.create_table('turn_feedback',
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('chat_id', sa.String(length=24), nullable=False),
+    sa.Column('turn_id', sa.String(length=256), nullable=False),
+    sa.Column('type', sa.String(length=50), nullable=False),
+    sa.Column('tag', sa.String(length=50), nullable=True),
+    sa.Column('message', sa.Text(), nullable=True),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('chat_id', 'turn_id')
+    )
+    op.create_index(op.f('ix_turn_feedback_user'), 'turn_feedback', ['user'], unique=False)
+    op.create_table('user',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('username', sa.String(length=256), nullable=True),
+    sa.Column('email', sa.String(length=254), nullable=True),
+    sa.Column('role', sa.String(length=50), nullable=False),
+    sa.Column('hashed_password', sa.String(length=128), nullable=False),
+    sa.Column('is_active', sa.Boolean(), nullable=False),
+    sa.Column('is_superuser', sa.Boolean(), nullable=False),
+    sa.Column('is_verified', sa.Boolean(), nullable=False),
+    sa.Column('is_staff', sa.Boolean(), nullable=False),
+    sa.Column('chat_collection_id', sa.String(length=24), nullable=True),
+    sa.Column('date_joined', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email'),
+    sa.UniqueConstraint('username')
+    )
+    op.create_index(op.f('ix_user_chat_collection_id'), 'user', ['chat_collection_id'], unique=False)
+    op.create_table('user_collection_subscription',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user_id', sa.String(length=24), nullable=False),
+    sa.Column('collection_marketplace_id', sa.String(length=24), nullable=False),
+    sa.Column('gmt_subscribed', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('user_id', 'collection_marketplace_id', 'gmt_deleted', name='idx_user_marketplace_history_unique')
+    )
+    op.create_index('idx_user_subscription_gmt_deleted', 'user_collection_subscription', ['gmt_deleted'], unique=False)
+    op.create_index('idx_user_subscription_marketplace', 'user_collection_subscription', ['collection_marketplace_id'], unique=False)
+    op.create_index('idx_user_subscription_user', 'user_collection_subscription', ['user_id'], unique=False)
+    op.create_table('user_quota',
+    sa.Column('user', sa.String(length=256), nullable=False),
+    sa.Column('key', sa.String(length=256), nullable=False),
+    sa.Column('quota_limit', sa.Integer(), nullable=False),
+    sa.Column('current_usage', sa.Integer(), nullable=False),
+    sa.Column('gmt_created', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_updated', sa.DateTime(timezone=True), nullable=False),
+    sa.Column('gmt_deleted', sa.DateTime(timezone=True), nullable=True),
+    sa.PrimaryKeyConstraint('user', 'key')
+    )
+    op.create_table('oauth_account',
+    sa.Column('id', sa.String(length=24), nullable=False),
+    sa.Column('user_id', sa.String(), nullable=False),
+    sa.Column('oauth_name', sa.String(length=100), nullable=False),
+    sa.Column('access_token', sa.String(length=1024), nullable=False),
+    sa.Column('expires_at', sa.Integer(), nullable=True),
+    sa.Column('refresh_token', sa.String(length=1024), nullable=True),
+    sa.Column('account_id', sa.String(length=320), nullable=False),
+    sa.Column('account_email', sa.String(length=320), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ondelete='cascade'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_oauth_account_account_id'), 'oauth_account', ['account_id'], unique=False)
+    op.create_index(op.f('ix_oauth_account_oauth_name'), 'oauth_account', ['oauth_name'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_oauth_account_oauth_name'), table_name='oauth_account')
+    op.drop_index(op.f('ix_oauth_account_account_id'), table_name='oauth_account')
+    op.drop_table('oauth_account')
+    op.drop_table('user_quota')
+    op.drop_index('idx_user_subscription_user', table_name='user_collection_subscription')
+    op.drop_index('idx_user_subscription_marketplace', table_name='user_collection_subscription')
+    op.drop_index('idx_user_subscription_gmt_deleted', table_name='user_collection_subscription')
+    op.drop_table('user_collection_subscription')
+    op.drop_index(op.f('ix_user_chat_collection_id'), table_name='user')
+    op.drop_table('user')
+    op.drop_index(op.f('ix_turn_feedback_user'), table_name='turn_feedback')
+    op.drop_table('turn_feedback')
+    op.drop_table('setting')
+    op.drop_index(op.f('ix_searchhistory_user'), table_name='searchhistory')
+    op.drop_index(op.f('ix_searchhistory_gmt_deleted'), table_name='searchhistory')
+    op.drop_index(op.f('ix_searchhistory_collection_id'), table_name='searchhistory')
+    op.drop_table('searchhistory')
+    op.drop_index('idx_questions_question_set_id', table_name='questions')
+    op.drop_table('questions')
+    op.drop_index('idx_question_sets_user_id', table_name='question_sets')
+    op.drop_index('idx_question_sets_collection_id', table_name='question_sets')
+    op.drop_table('question_sets')
+    op.drop_index(op.f('ix_prompt_template_user_id'), table_name='prompt_template')
+    op.drop_index(op.f('ix_prompt_template_scope'), table_name='prompt_template')
+    op.drop_index(op.f('ix_prompt_template_prompt_type'), table_name='prompt_template')
+    op.drop_index(op.f('ix_prompt_template_gmt_deleted'), table_name='prompt_template')
+    op.drop_table('prompt_template')
+    op.drop_index(op.f('ix_model_service_provider_status'), table_name='model_service_provider')
+    op.drop_index(op.f('ix_model_service_provider_name'), table_name='model_service_provider')
+    op.drop_index(op.f('ix_model_service_provider_gmt_deleted'), table_name='model_service_provider')
+    op.drop_table('model_service_provider')
+    op.drop_table('llm_provider_models')
+    op.drop_index(op.f('ix_llm_provider_user_id'), table_name='llm_provider')
+    op.drop_table('llm_provider')
+    op.drop_table('invitation')
+    op.drop_index('idx_graphindex_nodes_collection_type', table_name='graphindex_nodes')
+    op.drop_index('idx_graphindex_nodes_collection_name', table_name='graphindex_nodes')
+    op.drop_index('idx_graphindex_nodes_chunks', table_name='graphindex_nodes', postgresql_using='gin')
+    op.drop_table('graphindex_nodes')
+    op.drop_index('idx_graphindex_edges_collection_tgt', table_name='graphindex_edges')
+    op.drop_index('idx_graphindex_edges_collection_src', table_name='graphindex_edges')
+    op.drop_index('idx_graphindex_edges_chunks', table_name='graphindex_edges', postgresql_using='gin')
+    op.drop_table('graphindex_edges')
+    op.drop_index('idx_graphindex_chunks_collection_doc', table_name='graphindex_chunks')
+    op.drop_table('graphindex_chunks')
+    op.drop_index('idx_graph_curation_suggestions_run', table_name='graph_curation_suggestions')
+    op.drop_index('idx_graph_curation_suggestions_collection_status', table_name='graph_curation_suggestions')
+    op.drop_index('idx_graph_curation_suggestions_collection_created', table_name='graph_curation_suggestions')
+    op.drop_table('graph_curation_suggestions')
+    op.drop_index('idx_graph_curation_runs_user_created', table_name='graph_curation_runs')
+    op.drop_index('idx_graph_curation_runs_collection_status', table_name='graph_curation_runs')
+    op.drop_table('graph_curation_runs')
+    op.drop_index(op.f('ix_export_task_user'), table_name='export_task')
+    op.drop_index(op.f('ix_export_task_collection_id'), table_name='export_task')
+    op.drop_index('idx_export_task_user_status', table_name='export_task')
+    op.drop_index('idx_export_task_expires', table_name='export_task')
+    op.drop_table('export_task')
+    op.drop_index('idx_evaluations_user_id', table_name='evaluations')
+    op.drop_index('idx_evaluations_status', table_name='evaluations')
+    op.drop_index('idx_evaluations_collection_id', table_name='evaluations')
+    op.drop_table('evaluations')
+    op.drop_index('idx_evaluation_runs_user', table_name='evaluation_runs')
+    op.drop_index('idx_evaluation_runs_status', table_name='evaluation_runs')
+    op.drop_index('idx_evaluation_runs_dataset', table_name='evaluation_runs')
+    op.drop_index('idx_evaluation_runs_collection', table_name='evaluation_runs')
+    op.drop_index('idx_evaluation_runs_bot', table_name='evaluation_runs')
+    op.drop_table('evaluation_runs')
+    op.drop_index('idx_evaluation_run_items_status', table_name='evaluation_run_items')
+    op.drop_index('idx_evaluation_run_items_run', table_name='evaluation_run_items')
+    op.drop_table('evaluation_run_items')
+    op.drop_index('idx_evaluation_run_item_attempts_run', table_name='evaluation_run_item_attempts')
+    op.drop_index('idx_evaluation_run_item_attempts_item', table_name='evaluation_run_item_attempts')
+    op.drop_table('evaluation_run_item_attempts')
+    op.drop_index(op.f('ix_evaluation_items_status'), table_name='evaluation_items')
+    op.drop_index('idx_evaluation_items_evaluation_id', table_name='evaluation_items')
+    op.drop_table('evaluation_items')
+    op.drop_index(op.f('ix_evaluation_datasets_gmt_deleted'), table_name='evaluation_datasets')
+    op.drop_index('idx_evaluation_datasets_user', table_name='evaluation_datasets')
+    op.drop_index('idx_evaluation_datasets_collection', table_name='evaluation_datasets')
+    op.drop_table('evaluation_datasets')
+    op.drop_index(op.f('ix_evaluation_dataset_items_gmt_deleted'), table_name='evaluation_dataset_items')
+    op.drop_index('idx_evaluation_dataset_items_dataset', table_name='evaluation_dataset_items')
+    op.drop_table('evaluation_dataset_items')
+    op.drop_index(op.f('ix_document_index_status'), table_name='document_index')
+    op.drop_index(op.f('ix_document_index_index_type'), table_name='document_index')
+    op.drop_index(op.f('ix_document_index_id'), table_name='document_index')
+    op.drop_index(op.f('ix_document_index_document_id'), table_name='document_index')
+    op.drop_index('idx_document_index_status_lease', table_name='document_index')
+    op.drop_table('document_index')
+    op.drop_index('uq_document_collection_name_active', table_name='document', postgresql_where=sa.text('gmt_deleted IS NULL'))
+    op.drop_index(op.f('ix_document_user'), table_name='document')
+    op.drop_index(op.f('ix_document_status'), table_name='document')
+    op.drop_index(op.f('ix_document_gmt_deleted'), table_name='document')
+    op.drop_index(op.f('ix_document_content_hash'), table_name='document')
+    op.drop_index(op.f('ix_document_collection_id'), table_name='document')
+    op.drop_table('document')
+    op.drop_table('config')
+    op.drop_index(op.f('ix_collection_summary_status'), table_name='collection_summary')
+    op.drop_index(op.f('ix_collection_summary_collection_id'), table_name='collection_summary')
+    op.drop_index('idx_collection_summary_status_lease', table_name='collection_summary')
+    op.drop_table('collection_summary')
+    op.drop_index('idx_collection_marketplace_status', table_name='collection_marketplace')
+    op.drop_index('idx_collection_marketplace_list', table_name='collection_marketplace')
+    op.drop_index('idx_collection_marketplace_gmt_deleted', table_name='collection_marketplace')
+    op.drop_index('idx_collection_marketplace_collection_id', table_name='collection_marketplace')
+    op.drop_table('collection_marketplace')
+    op.drop_index(op.f('ix_collection_user'), table_name='collection')
+    op.drop_index(op.f('ix_collection_status'), table_name='collection')
+    op.drop_index(op.f('ix_collection_gmt_deleted'), table_name='collection')
+    op.drop_table('collection')
+    op.drop_index(op.f('ix_chat_user'), table_name='chat')
+    op.drop_index(op.f('ix_chat_status'), table_name='chat')
+    op.drop_index(op.f('ix_chat_gmt_deleted'), table_name='chat')
+    op.drop_index(op.f('ix_chat_bot_id'), table_name='chat')
+    op.drop_table('chat')
+    op.drop_index(op.f('ix_bot_user'), table_name='bot')
+    op.drop_index(op.f('ix_bot_status'), table_name='bot')
+    op.drop_index(op.f('ix_bot_gmt_deleted'), table_name='bot')
+    op.drop_table('bot')
+    op.drop_index('idx_audit_user_id', table_name='audit_log')
+    op.drop_index('idx_audit_status_code', table_name='audit_log')
+    op.drop_index('idx_audit_start_time', table_name='audit_log')
+    op.drop_index('idx_audit_resource_type', table_name='audit_log')
+    op.drop_index('idx_audit_resource_id', table_name='audit_log')
+    op.drop_index('idx_audit_request_id', table_name='audit_log')
+    op.drop_index('idx_audit_http_method', table_name='audit_log')
+    op.drop_index('idx_audit_gmt_created', table_name='audit_log')
+    op.drop_index('idx_audit_api_name', table_name='audit_log')
+    op.drop_table('audit_log')
+    op.drop_index(op.f('ix_api_key_user'), table_name='api_key')
+    op.drop_index(op.f('ix_api_key_status'), table_name='api_key')
+    op.drop_index(op.f('ix_api_key_is_system'), table_name='api_key')
+    op.drop_index(op.f('ix_api_key_gmt_deleted'), table_name='api_key')
+    op.drop_table('api_key')
+    op.drop_index(op.f('ix_agent_turn_user'), table_name='agent_turn')
+    op.drop_index(op.f('ix_agent_turn_status'), table_name='agent_turn')
+    op.drop_index(op.f('ix_agent_turn_request_id'), table_name='agent_turn')
+    op.drop_index(op.f('ix_agent_turn_reference_bundle_artifact_id'), table_name='agent_turn')
+    op.drop_index(op.f('ix_agent_turn_chat_id'), table_name='agent_turn')
+    op.drop_index(op.f('ix_agent_turn_bot_id'), table_name='agent_turn')
+    op.drop_index(op.f('ix_agent_turn_answer_artifact_id'), table_name='agent_turn')
+    op.drop_index('idx_agent_turn_user_status', table_name='agent_turn')
+    op.drop_index('idx_agent_turn_chat_created', table_name='agent_turn')
+    op.drop_table('agent_turn')
+    op.drop_index(op.f('ix_agent_timeline_event_type'), table_name='agent_timeline_event')
+    op.drop_index(op.f('ix_agent_timeline_event_turn_id'), table_name='agent_timeline_event')
+    op.drop_index(op.f('ix_agent_timeline_event_timestamp'), table_name='agent_timeline_event')
+    op.drop_index('idx_agent_timeline_event_turn_timestamp', table_name='agent_timeline_event')
+    op.drop_table('agent_timeline_event')
+    op.drop_index(op.f('ix_agent_artifact_turn_id'), table_name='agent_artifact')
+    op.drop_index(op.f('ix_agent_artifact_artifact_type'), table_name='agent_artifact')
+    op.drop_index('idx_agent_artifact_turn_type', table_name='agent_artifact')
+    op.drop_table('agent_artifact')
+    # ### end Alembic commands ###
